server/requirements.txt
fastapi
uvicorn[standard]
pydantic
llama-cpp-python # For LLM interaction
# llama-cpp-python # Add later in P4 